+++ eRestaurant/Controllers/Master/RawMaterialController.cs
@@ -12,21 +12,27 @@ namespace RocketPOS.Controllers.Master
 {
     public class RawMaterialController : Controller
     {
+        private readonly ICommonService _iCommonService;
         private readonly IRawMaterialService _iRawMaterialService;
         private IStringLocalizer<RocketPOSResources> _sharedLocalizer;
         private LocService _locService;
 
-        public RawMaterialController(IRawMaterialService iRawMaterialService, IStringLocalizer<RocketPOSResources> sharedLocalizer, LocService locService)
+        public RawMaterialController(IRawMaterialService iRawMaterialService, ICommonService iCommonService, IStringLocalizer<RocketPOSResources> sharedLocalizer, LocService locService)
         {
             _iRawMaterialService = iRawMaterialService;
+            _iCommonService = iCommonService;
             _sharedLocalizer = sharedLocalizer;
             _locService = locService;
         }
 
-        public ActionResult Index()
+        public ActionResult Index(int? noDelete)
         {
             List<RawMaterialModel> rawMaterialModel = new List<RawMaterialModel>();
             rawMaterialModel = _iRawMaterialService.GetRawMaterialList().ToList();
+            if (noDelete != null)
+            {
+                ViewBag.Result = _locService.GetLocalizedHtmlString("Can not delete reference available.");
+            }
             return View(rawMaterialModel);
         }
 
@@ -82,8 +88,17 @@ namespace RocketPOS.Controllers.Master
 
         public ActionResult Delete(int id)
         {
-            var deletedid = _iRawMaterialService.DeleteRawMaterial(id);
-            return RedirectToAction(nameof(Index));
+            int result = 0;
+            result = _iCommonService.GetValidateReference("RawMaterial", id.ToString());
+            if (result > 0)
+            {
+                return RedirectToAction(nameof(Index), new { noDelete = result });
+            }
+            else
+            {
+                var deletedid = _iRawMaterialService.DeleteRawMaterial(id);
+                return RedirectToAction(nameof(Index));
+            }
         }
 
         private string ValidationRawMaterial(RawMaterialModel rawMaterialModel)
